Refactor: 가독성 향상 및 원할한 테스트를 위해 카프카 이벤트 생성 및 로그 기록 코드 리팩토링
